C03_Newweb/PIVOT/pivot_back/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
python-dotenv
oracledb
